Move GlobalConfig to dedicated package
