+++ core/src/main/java/org/dockbox/darwin/core/server/config/DefaultGlobalConfig.java
@@ -1,8 +1,10 @@
-package org.dockbox.darwin.core.server;
+package org.dockbox.darwin.core.server.config;
 
 import com.google.inject.Inject;
 
 import org.dockbox.darwin.core.i18n.Languages;
+import org.dockbox.darwin.core.server.Server;
+import org.dockbox.darwin.core.server.config.GlobalConfig;
 import org.dockbox.darwin.core.util.files.ConfigManager;
 import org.jetbrains.annotations.NotNull;
 
@@ -22,6 +24,23 @@ public class DefaultGlobalConfig implements GlobalConfig {
     public Languages getDefaultLanguage() {
         String cfg = getSetting(ConfigKeys.GLOBAL_CONFIG.getKey());
         return cfg == null ? Languages.EN_US : Languages.valueOf(cfg.toUpperCase());
+    }
+
+    @Override
+    public boolean getStacktracesAllowed() {
+        String cfg = getSetting(ConfigKeys.ALLOW_STACKTRACES.getKey());
+        return Boolean.parseBoolean(cfg);
+    }
+
+    @NotNull
+    @Override
+    public ExceptionLevels getExceptionLevel() {
+        String cfg = getSetting(ConfigKeys.EXCEPTION_LEVEL.getKey());
+        try {
+            return ExceptionLevels.valueOf(cfg.toUpperCase());
+        } catch (NullPointerException | IllegalArgumentException e) {
+            return ExceptionLevels.MINIMAL;
+        }
     }
 
     @Override
