+++ core/src/main/kotlin/org/dockbox/darwin/core/server/KServer.kt
@@ -1,5 +1,6 @@
 package org.dockbox.darwin.core.server
 
+import org.dockbox.darwin.core.server.config.GlobalConfig
 import org.slf4j.Logger
 import java.util.*
